登录3——创建实体类模块

+++ LZMall-parent/lzm-common/src/main/java/org/lzmcommon/entity/BaseEntity.java
@@ -3,12 +3,13 @@ package org.lzmcommon.entity;
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @Data
 public class BaseEntity {
     @TableId(type = IdType.AUTO)
-    private long id;
+    private BigInteger id;
     @TableField(fill = FieldFill.INSERT)
     private Timestamp createTime;
     @TableField(fill = FieldFill.INSERT_UPDATE)
